refactor: split grade formula params store into data layer
Also makes the params path env-overridable (QUODEQ_GRADE_FORMULA_PATH) and
sandboxes it in the test conftest: two rescore-path tests failed on any
branch the moment the developer's real ~/.quodeq/grade_formula.json held
custom params.
