+++ src/quodeq/data/projection/grade_projector.py
@@ -14,6 +14,7 @@ from pathlib import Path
 
 from quodeq.core.scoring.params import ScoringParams
 from quodeq.core.types.finding import Finding
+from quodeq.data.fs.grade_formula_store import load_params
 from quodeq.data.sqlite._row_mappers import row_to_finding
 from quodeq.data.sqlite.connection import open_evaluation_db
 from quodeq.data.sqlite.state_store import SQLiteStateStore
@@ -129,8 +130,7 @@ def recompute_grades(run_dir: Path, params: ScoringParams | None = None) -> None
     When *params* is None, the saved grade-formula params are loaded.
     """
     if params is None:
-        from quodeq.services import grade_formula  # noqa: PLC0415
-        params = grade_formula.load_params()
+        params = load_params()
     principle_rows, dimension_rows = compute_run_grades(run_dir, params)
 
     # Carry the per-dim exit_reason (failure_streak, time_limit, ...) from the
