+++ src/quodeq/services/grade_formula.py
@@ -1,69 +1,28 @@
-"""Persistence for user-tuned grade formula parameters.
+"""User-tuned grade formula: apply/preview orchestration.
 
-The file at ``~/.quodeq/grade_formula.json`` holds the camelCase dict shape
-from ``params_to_dict``. Absent file means Q² defaults. A corrupt file logs
-a warning and falls back to defaults rather than breaking every score read.
+The parameter file itself is owned by ``data/fs/grade_formula_store.py``;
+the accessors are re-exported here because the API layer talks to services,
+not to data.
 """
 from __future__ import annotations
 
-import json
 import logging
 import time
 from dataclasses import dataclass
 from pathlib import Path
 
-from quodeq.core.scoring.params import (
-    DEFAULT_PARAMS,
-    ScoringParams,
-    params_from_dict,
-    params_to_dict,
-    validate_params,
+from quodeq.core.scoring.params import ScoringParams
+from quodeq.core.scoring.projector_scoring import compute_run_score
+from quodeq.data.fs.grade_formula_store import (  # noqa: F401 — re-exported API
+    grade_formula_path,
+    is_custom,
+    load_params,
+    reset_params,
+    save_params,
 )
 from quodeq.shared.run_status import UnsupportedSchemaError, read_status
 
 _logger = logging.getLogger(__name__)
-
-
-def grade_formula_path() -> Path:
-    """Location of the custom-params file (function so tests can monkeypatch)."""
-    return Path.home() / ".quodeq" / "grade_formula.json"
-
-
-def load_params() -> ScoringParams:
-    """Return saved custom params, or Q² defaults when absent or unreadable."""
-    path = grade_formula_path()
-    if not path.is_file():
-        return DEFAULT_PARAMS
-    try:
-        data = json.loads(path.read_text(encoding="utf-8"))
-        params = params_from_dict(data)
-    except (OSError, json.JSONDecodeError, AttributeError, KeyError, TypeError, ValueError) as exc:
-        _logger.warning("Unreadable %s (%s); using Q2 default formula.", path, exc)
-        return DEFAULT_PARAMS
-    if validate_params(params):
-        _logger.warning("Invalid params in %s; using Q2 default formula.", path)
-        return DEFAULT_PARAMS
-    return params
-
-
-def save_params(params: ScoringParams) -> None:
-    """Validate and persist custom params. Raises ValueError when invalid."""
-    errors = validate_params(params)
-    if errors:
-        raise ValueError("; ".join(errors))
-    path = grade_formula_path()
-    path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_text(json.dumps(params_to_dict(params), indent=2), encoding="utf-8")
-
-
-def reset_params() -> None:
-    """Remove the custom-params file (back to Q² defaults)."""
-    grade_formula_path().unlink(missing_ok=True)
-
-
-def is_custom() -> bool:
-    """True when a custom-params file is in effect."""
-    return grade_formula_path().is_file()
 
 
 def _run_recency_key(run_dir: Path) -> tuple[int, str | float]:
@@ -174,7 +133,6 @@ def preview_scores(
     """
     from quodeq.data.projection.grade_projector import compute_run_grades  # noqa: PLC0415
     from quodeq.data.sqlite.state_store import SQLiteStateStore  # noqa: PLC0415
-    from quodeq.core.scoring.projector_scoring import compute_run_score  # noqa: PLC0415
 
     project_dir = reports_root / project
     if not project_dir.is_dir():
