+++ tests/api/test_grade_formula_api.py
@@ -24,7 +24,7 @@ def _disable_auth(monkeypatch):
 @pytest.fixture
 def formula_path(tmp_path, monkeypatch):
     path = tmp_path / "grade_formula.json"
-    monkeypatch.setattr(grade_formula, "grade_formula_path", lambda: path)
+    monkeypatch.setenv("QUODEQ_GRADE_FORMULA_PATH", str(path))
     return path
