+++ tests/services/test_grade_formula.py
@@ -12,7 +12,7 @@ from quodeq.services import grade_formula
 @pytest.fixture
 def formula_path(tmp_path, monkeypatch):
     path = tmp_path / "grade_formula.json"
-    monkeypatch.setattr(grade_formula, "grade_formula_path", lambda: path)
+    monkeypatch.setenv("QUODEQ_GRADE_FORMULA_PATH", str(path))
     return path
